+++ src/components/signin/index.ts
@@ -1,6 +1,7 @@
 export { default as Thumbnail } from "./Thumbnail";
 export { default as SigninHeader } from "./Header";
 export { default as SigninForm } from "./SigninForm";
+export { default as SigninLayout } from "./SigninLayout";
 export { default as UserTypeTag } from "./UserTypeTag";
 
 export { default as CreateAdminButton } from "./admin/CreateAdminButton";
